Ditance sensors start

+++ RacerbilsEvolutionsProgram/CarController.pde
@@ -19,9 +19,9 @@ class CarController implements Comparable <CarController>{
     sensorSystem.updateSensorsignals(bil.pos, bil.vel);
     //3.)hjernen beregner hvor meget der skal drejes
     float turnAngle = 0;
-    float x1 = (float)int(sensorSystem.leftSensorSignal);
-    float x2 = (float)int(sensorSystem.frontSensorSignal);
-    float x3 = (float)int(sensorSystem.rightSensorSignal);    
+    float x1 = sensorSystem.leftSensorSignal;
+    float x2 = sensorSystem.frontSensorSignal;
+    float x3 = sensorSystem.rightSensorSignal;    
     turnAngle = hjerne.getOutput(x1, x2, x3);    
     //4.)bilen drejes
     bil.turnCar(turnAngle);
